Move reports views, forms and tests to `reports` application
